refactor button input to work based on milliseconds elapsed instead of tickCount

+++ BulbChipSTM32C071FBPx/Core/Inc/chip_state.h
@@ -15,7 +15,7 @@
 #include "mc3479.h"
 #include "rgb.h"
 
-typedef float MillisForElapsedChipTicks(uint16_t elapsedTicks);
+typedef uint16_t MillisForElapsedChipTicks(uint16_t elapsedTicks);
 
 typedef void WriteToUsbSerial(uint8_t itf, const char *buf, uint32_t count);
